requirements.txt
python=3.7.3
pytorch=1.2.0
torchvision=0.4.0
numpy=1.16.2
tensorboardx=1.9
matplotlib=3.0.3
tqdm=4.31.1
prettytable=0.7.2
